pyproject.toml
[project]
name = "eleanor"
version = "0.0.0"
authors = [
  {name = "39 Alpha Research, Team 0", email = "39alpha@39alpharesearch.org"},
]
description = "Large-scale, aqueous geochemical modeling"
readme = "README.rst"
license = {file = "LICENSE"}
dependencies = [
  'antlr4-python3-runtime',
  'matplotlib',
  'numpy',
  'pandas',
  'pyyaml',
  'tqdm',
]
requires-python = ">= 3.12"

[project.urls]
Repository = "https://github.com/39alpha/eleanor.git"
Issues = "https://github.com/39alpha/eleanor/issues"
Changelog = "https://github.com/39alpha/eleanor/blob/master/CHANGELOG.md"

[project.scripts]
eleanor = "eleanor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.yapf]
column_limit = 120

[tool.isort]
line_length = 120

[tool.mypy]
check_untyped_defs = true
exclude = "(docs/source/conf\\.py|eleanor/hanger/eq36/parsers/.*\\.py)"
